# api/api-docs/build/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Xpense web api
  description: Xpense web api.
servers:
  - url: 'http://localhost:7000/api'
  - url: 'https://dev.api.xpense.bulldozerinc.com/api'
paths:
...
  '/org/{orgId}/sheets':
    get:
      summary: Get user organization sheets
      tags:
        - Organization
      operationId: getSheets
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Current user organization sheets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sheet'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /users:
    get:
      summary: List of filtered users by organization
      tags:
        - Users
      operationId: usersBYorganization
      parameters:
        - in: query
          name: orgId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: List of filtered users by organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  '/users/{userId}':
    patch:
      summary: Change user role
      tags:
        - Users
      operationId: changeUserRole
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
      requestBody:
        description: Role ID
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleId:
                  type: string
      responses:
        '204':
          description: Role successfully associated with user.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  '/users/{userId}/org/{orgId}':
    patch:
      summary: Remove a user from an orgnization
      tags:
        - Users
      operationId: removeUserFromOrg
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Role successfully associated with user.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /roles:
    get:
      summary: List of roles filtred by name
      tags:
        - Users
      operationId: getRoles
      parameters:
        - in: query
          name: name
          schema:
            type: string
          required: false
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Roles'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Profile:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        name:
          type: string
        nickname:
          type: string
        lastName:
          type: string
        firstName:
          type: string
        picture:
          type: string
        defaultCurrencyId:
          type: string
        phoneNumber:
          type: string
        organizations:
          type: array
          items:
            type: string
...
    Sheet:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        status:
          type: string
        creationDate:
          type: string
          format: date-time
        submittedOn:
          type: string
          format: date-time
        exported:
          type: boolean
        total:
          type: integer
...
    Users:
      type: array
      items:
        $ref: '#/components/schemas/Profile'
    Roles:
      type: array
      items:
        type: object
        properties:
          id:
            type: string
          name:
            type: string
          description:
            type: string
            nullable: true
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
    NotFound:
      description: The specified resource was not found
    BadRequest:
      description: Bad request
security:
  - bearerAuth: []